Refactor: Refinements and code cleanup for chatbot
