+++ chatbot/chatbot_utilities.py
@@ -1,5 +1,6 @@
 from dataclasses import dataclass, field
 from typing import Any
+from enum import Enum
 
 
 class ResponseError(ValueError):
@@ -12,3 +13,10 @@ class ResponseError(ValueError):
 class Response:
     raw_response: str
     processed_response: Any
+
+class ChatbotState(Enum):
+    BEGINNING = 1
+    QUESTIONING = 2
+    REVIEWING = 3
+    MODIFYING_SELECTION = 4
+    MODIFYING = 5
